refactor: test 파일 패키지화
