+++ src/test/java/test/linegenerator/SuccessiveLineGenerator.java
@@ -1,8 +1,10 @@
-package ladder.domain;
+package test.linegenerator;
+
+import ladder.domain.LineGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
-//TODO 패키지 이동(test와 구분)
+
 public class SuccessiveLineGenerator implements LineGenerator {
     @Override
     public List<Boolean> getLine(int peopleNumber) {
